GROUNDWORK-491-serialization-wip:  saving WIP snapshot as backup
This round adds a bunch of test cases (a few of which are skipped
at the moment, as the routines they call are not completely
implemented).  But almost everything which is enabled is working,
based solely on generated code and no longer on any of the early
manually-written template code, and most tests pass.  An exception
is suppression of a zero time value on JSON output, which is not
yet implemented because it has to be special-cased in the code.
There are some other places as well where omitempty is not yet
supported.