GROUNDWORK-491-serialization-wip:  snapshot of WIP for backup purposes
This snapshot finds the set of unit tests extended and all passing, to
successfully cover the ground that they do.  Still remaining is work to
verify that omitempty and alternate field names in struct field tags
are implemented for every last datatype.  There is also further code
cleanup to do, and the task list will need review to verify that all
the items it mentions are covered.