GROUNDWORK-491-serialization-wip:  snapshot of WIP for backup purposes
This iteration applies the first rounds of code cleanup.  (There will be
more rounds in the future, as we use the code in practice and as we feel
comfortable with deleting more of the scaffolding we used in the process
of building the working code.)  Much of the original generated code that
we won't be using in practice is no longer generated.