GROUNDWORK-491-serialization-wip:  snapshot of WIP for backup purposes
This snapshot captures what appears to be fully working conversions
between JSON strings and C structures, in both directions, without
any memory leaks.  The generated code is fully passing the unit
testing when run against the full set of JSON strings we have so
far, corresponding to the test copies of transit.go and related
files in the tool source directory.  The gotocjson.go code generator
has had a fair amount of cleanup with regard to scaffolding put in
place during development, but more remains to be done in that area.
There are also still code generation improvements of various kinds
to be made.  But this is at least a working copy of the development
stream, as far as it goes.