GROUNDWORK-787-topological-sorting:  fix topological sorting of Go datatypes
Simple typedefs which were not in fact "simple", but involved some punctuation
because they were more-complex constructions such as "type FooBars []*FooBar",
were not being correctly handled during topological sorting.  The result was
that the generated C code could contain forward references, which would fail
to compile.  The fix in the conversion tool is to extract the underlying Go
datatype from the typedef and use that instead of the C-language datatype,
when we detect a complex simple-typedef construction.  This allows the actual
datatype dependencies to factor into the sorting.