GROUNDWORK-770-TimeValue-fix:  fix omitempty code for a ptr to a struct
This change will improve the code that handles a pointer to a struct, so it correctly
understands a NULL pointer to such a struct and does not attempt to dive deeper while
converting a tree of Transit structures into JSON.  This will remove the unnecessary
messages we were seeing in a debug session regarding millisecondTimestamp conversions.
Those messages were coming from TypedValue structures that did not define the TimeValue
member, which is now a pointer instead of an embedded struct.