GROUNDWORK-770-ignore-interfaces:  ignore struct fields using "interface"
We implement a somewhat sneaky workaround for a Go map[string]interface{} field in a
structure, by simply suppressing that entire field in the corresponding C structure.
That should be adequate for present purposes, because we do not expect to ever have
a need to pass that structure (MonitorConnection) between Go code and C code.

As a side effect that won't affect any current generated code, but as a general code
improvement, the gotocjson tool also now will simply ignore structure fields that start
with a non-uppercase letter, in the same way that Go implements its own field-visibility
rules for structure fields.