GROUNDWORK-1066-extension:  gotocjson must factor out generic datatypes
The gotocjson tool has been modified to process a new generic_datatypes.go
file that we can modify to contain complex generic datatypes such as
map[string]string.  Support for such datatypes in the generated C code
needs to be factored out and not included in every converted application
Go packge that needs such datatypes, lest there be conflicts in generated
C header and object files.

With this change, application Go packages can use the map[string]string
datatype as the type of a structure field, without any attempt to apply
any obscuring type definitions to attempt to evade the limitations of
the conversion tool before this extension.

The TCG Makefile's have beee modified to run the gotocjson tool with the
proper new option (-g, for "generic") to process the generic_datatypes.go
file in a special mode, wherein the tool understands that in this case,
the complex generic datatypes should be generated instead of suppressed.
The resulting libtransitjson.so library should link with DataGeyser
without problem.