GROUNDWORK-1305/feature:  change character encoding of Nagios data
Support transforming ISO-8859-1 string data obtained from Nagios
to UTF-8 encoding as needed in the JSON we pass to TCG.

This implementation handles all the C-to-Go transforms, and prepares
for eventual Go-to-C transforms, but does not yet make any of the
latter.  Those are not at issue at the moment, and we can add that
capability at a future time.