GROUNDWORK-1337-feature:  support "type foos []foo" in gotocjson (#199)
* GROUNDWORK-1337-feature:  support "type foos []foo" in gotocjson

This commit:
* adds gotocjson support for declaring a simple type as an array of other types
* renames HostServiceInDowntimeList as its proper plural (HostServicesInDowntime)
* adds "omitempty" to the json tag for HostServiceInDowntime.ServiceDescription

* GROUNDWORK-1337-feature:  revise handling of an array-typedef

Revise the C-routine declarations and #define macros emitted when
handling simple typedefs.  This provides the full set of resources
needed by DataGeyser.

Do a better job of aligning generated code, for clarity when people
read that code.

* GROUNDWORK-1337-feature:  add missing generated C declarations

This commit adds some missing declarations generated in C code
from Go code, and cleans up some of the other declarations.  The
result is now sufficient to meet the updated needs of DataGeyser.

* GROUNDWORK-1337-feature:  update HostServicesInDowntime type definition to match the Foundation API

Co-authored-by: Glenn Herteg <gherteg@digit7.groundwork.groundworkopensource.com>
Co-authored-by: Pavlo Sumkin <pavlo@bluesunrise.com>