GW-1376-time-duration:  extend gotocjson to handle time.Duration fields
This commit allows Go struct fields to be declared as type time.Duration,
while generating correct conversion code in that case.  The tool has been
extended past that point in pursuit of processing the entire official
/usr/local/go/src/time/time.go file, but ultimately that was not possible.
Instead, we use a highly-abridged copy of that file, and that suffices
for current use against our application code.