GROUNDWORK-1674-add-perl:  add Perl and xorp.pl to the tcg container
We have found that our SNMPv3 code currently calls a small Perl helper
script to do its work.  In the interest of getting this working quickly
for a POC, I have modified the Dockerfile in a manner which I believe
will do the right things (adding Perl to the container, and installing
the Perl help script where it seems to be needed in actual usage).

Once we have validated that these changes have resolved the immediate
need, we can revisit these changes to see if a different permanent
solution is required.